api-gateway/pyproject.toml
# api-gateway/pyproject.toml
[tool.poetry]
name = "api-gateway"
version = "1.0.0"
description = "API Gateway for Nyro Microservices"
authors = ["Nyro <dev@nyro.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0" # Actualizado
uvicorn = {extras = ["standard"], version = "^0.28.0"} # Actualizado
gunicorn = "^21.2.0" # Para despliegue
pydantic = {extras = ["email"], version = "^2.6.4"} # Actualizado a v2
pydantic-settings = "^2.2.1" # Para configuración
httpx = "^0.27.0" # Cliente HTTP async
python-jose = {extras = ["cryptography"], version = "^3.3.0"} # Para JWT
structlog = "^24.1.0" # Para logging estructurado
tenacity = "^8.2.3" # Para reintentos en httpx (opcional pero bueno)

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.21.1"
pytest-httpx = "^0.29.0" # Para mockear llamadas HTTP en tests

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"